create probe package
